+++ invenio_rdm_records/assets/semantic-ui/js/invenio_rdm_records/src/deposit/fields/FileUploader/FileUploader.js
@@ -1,5 +1,5 @@
 // This file is part of Invenio-RDM-Records
-// Copyright (C) 2020-2023 CERN.
+// Copyright (C) 2020-2025 CERN.
 // Copyright (C) 2020-2022 Northwestern University.
 // Copyright (C)      2022 Graz University of Technology.
 // Copyright (C)      2022 TU Wien.
@@ -195,6 +195,7 @@ export const FileUploaderComponent = ({
       warningMsg={warningMsg}
       setWarningMsg={setWarningMsg}
       filesLocked={lockFileUploader}
+      hasError={hasError}
       {...uiProps}
     >
       <>
@@ -252,6 +253,7 @@ export const FileUploaderComponent = ({
             filesList={filesList}
             dropzoneParams={dropzoneParams}
             filesLocked={lockFileUploader}
+            hasError={hasError}
             filesEnabled={filesEnabled}
             deleteFile={deleteFile}
             decimalSizeDisplay={decimalSizeDisplay}
@@ -264,6 +266,7 @@ export const FileUploaderComponent = ({
                   filesList={filesList}
                   dropzoneParams={dropzoneParams}
                   filesLocked={lockFileUploader}
+                  hasError={hasError}
                   filesEnabled={filesEnabled}
                   deleteFile={deleteFile}
                   decimalSizeDisplay={decimalSizeDisplay}